+++ src/models/general.rs
@@ -1,6 +1,7 @@
 use std::fmt::Debug;
 
 use chrono::NaiveDate;
+use rust_decimal::Decimal;
 use serde::Serialize;
 
 #[derive(Clone, Debug, Serialize)]
@@ -11,8 +12,8 @@ pub struct Detraccion {
 
     /// Catalog54
     pub tipo_bien_detraido: &'static str,
-    pub porcentaje: f64,
-    pub monto: Option<f64>,
+    pub porcentaje: Decimal,
+    pub monto: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Serialize)]
@@ -25,12 +26,12 @@ pub enum TipoFormaDePago {
 pub struct FormaDePago {
     pub tipo: TipoFormaDePago,
     pub cuotas: Vec<CuotaDePago>,
-    pub total: Option<f64>,
+    pub total: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct CuotaDePago {
-    pub importe: f64,
+    pub importe: Decimal,
     pub fecha_pago: NaiveDate,
 }
 
@@ -39,10 +40,10 @@ pub struct Percepcion {
     /// Catalog53
     pub tipo: &'static str,
 
-    pub porcentaje: Option<f64>,
-    pub monto: Option<f64>,
-    pub monto_base: Option<f64>,
-    pub monto_total: Option<f64>,
+    pub porcentaje: Option<Decimal>,
+    pub monto: Option<Decimal>,
+    pub monto_base: Option<Decimal>,
+    pub monto_total: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Serialize)]
@@ -56,48 +57,48 @@ pub struct Anticipo {
     /// Catalog12.
     pub comprobante_tipo: Option<&'static str>,
 
-    pub monto: f64,
+    pub monto: Decimal,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct Descuento {
     /// Catalog53.
     pub tipo: Option<&'static str>,
-    pub monto: f64,
-    pub monto_base: Option<f64>,
-    pub factor: Option<f64>,
+    pub monto: Decimal,
+    pub monto_base: Option<Decimal>,
+    pub factor: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct Guia {
     /// Catalog1. Valores válidos: "09", "31"
-    pub tipo_documento: f64,
+    pub tipo_documento: Decimal,
     pub serie_numero: &'static str,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct DocumentoRelacionado {
     /// Catalog12.
-    pub tipo_documento: f64,
+    pub tipo_documento: Decimal,
     pub serie_numero: &'static str,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct Detalle {
     pub descripcion: &'static str,
-    pub cantidad: f64,
+    pub cantidad: Decimal,
     pub unidad_medida: Option<&'static str>,
 
     /// Precio + bool. True si el precio incluye impuestos, false si no incluye impuestos
-    pub precio: Option<f64>,
-    pub precio_con_impuestos: Option<f64>,
-    pub precio_referencia: Option<f64>,
+    pub precio: Option<Decimal>,
+    pub precio_con_impuestos: Option<Decimal>,
+    pub precio_referencia: Option<Decimal>,
     /// Catalog16
     pub precio_referencia_tipo: Option<&'static str>,
 
-    pub igv_tasa: Option<f64>,
-    pub icb_tasa: Option<f64>,
-    pub isc_tasa: Option<f64>,
+    pub igv_tasa: Option<Decimal>,
+    pub icb_tasa: Option<Decimal>,
+    pub isc_tasa: Option<Decimal>,
 
     /// Catalog7
     pub igv_tipo: Option<&'static str>,
@@ -105,42 +106,42 @@ pub struct Detalle {
     pub isc_tipo: Option<&'static str>,
 
     pub icb_aplica: bool,
-    pub icb: Option<f64>,
+    pub icb: Option<Decimal>,
 
-    pub igv: Option<f64>,
-    pub igv_base_imponible: Option<f64>,
+    pub igv: Option<Decimal>,
+    pub igv_base_imponible: Option<Decimal>,
 
-    pub isc: Option<f64>,
-    pub isc_base_imponible: Option<f64>,
+    pub isc: Option<Decimal>,
+    pub isc_base_imponible: Option<Decimal>,
 
-    pub total_impuestos: Option<f64>,
+    pub total_impuestos: Option<Decimal>,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct TotalImporte {
-    pub anticipos: f64,
-    pub descuentos: f64,
-    pub importe: f64,
-    pub importe_sin_impuestos: f64,
-    pub importe_con_impuestos: f64,
+    pub anticipos: Decimal,
+    pub descuentos: Decimal,
+    pub importe: Decimal,
+    pub importe_sin_impuestos: Decimal,
+    pub importe_con_impuestos: Decimal,
 }
 
 #[derive(Clone, Debug, Serialize)]
 pub struct TotalImpuestos {
-    pub total: f64,
-    pub ivap_importe: f64,
-    pub ivap_base_imponible: f64,
-    pub exportacion_importe: f64,
-    pub exportacion_base_imponible: f64,
-    pub gravado_importe: f64,
-    pub gravado_base_imponible: f64,
-    pub inafecto_importe: f64,
-    pub inafecto_base_imponible: f64,
-    pub exonerado_importe: f64,
-    pub exonerado_base_imponible: f64,
-    pub gratuito_importe: f64,
-    pub gratuito_base_imponible: f64,
-    pub icb_importe: f64,
-    pub isc_importe: f64,
-    pub isc_base_imponible: f64,
+    pub total: Decimal,
+    pub ivap_importe: Decimal,
+    pub ivap_base_imponible: Decimal,
+    pub exportacion_importe: Decimal,
+    pub exportacion_base_imponible: Decimal,
+    pub gravado_importe: Decimal,
+    pub gravado_base_imponible: Decimal,
+    pub inafecto_importe: Decimal,
+    pub inafecto_base_imponible: Decimal,
+    pub exonerado_importe: Decimal,
+    pub exonerado_base_imponible: Decimal,
+    pub gratuito_importe: Decimal,
+    pub gratuito_base_imponible: Decimal,
+    pub icb_importe: Decimal,
+    pub isc_importe: Decimal,
+    pub isc_base_imponible: Decimal,
 }
