+++ src/models/traits/detalle/icbtasa.rs
@@ -1,21 +1,22 @@
 use crate::models::general::Detalle;
+use rust_decimal::Decimal;
 
 pub trait DetalleICBTasaGetter {
-    fn get_icbtasa(&self) -> &Option<f64>;
+    fn get_icbtasa(&self) -> &Option<Decimal>;
 }
 
 pub trait DetalleICBTasaSetter {
-    fn set_icbtasa(&mut self, val: f64);
+    fn set_icbtasa(&mut self, val: Decimal);
 }
 
 impl DetalleICBTasaGetter for Detalle {
-    fn get_icbtasa(&self) -> &Option<f64> {
+    fn get_icbtasa(&self) -> &Option<Decimal> {
         &self.icb_tasa
     }
 }
 
 impl DetalleICBTasaSetter for Detalle {
-    fn set_icbtasa(&mut self, val: f64) {
+    fn set_icbtasa(&mut self, val: Decimal) {
         self.icb_tasa = Some(val);
     }
 }
