+++ src/models/traits/detalle/igv.rs
@@ -1,21 +1,22 @@
 use crate::models::general::Detalle;
+use rust_decimal::Decimal;
 
 pub trait DetalleIGVGetter {
-    fn get_igv(&self) -> &Option<f64>;
+    fn get_igv(&self) -> &Option<Decimal>;
 }
 
 pub trait DetalleIGVSetter {
-    fn set_igv(&mut self, val: f64);
+    fn set_igv(&mut self, val: Decimal);
 }
 
 impl DetalleIGVGetter for Detalle {
-    fn get_igv(&self) -> &Option<f64> {
+    fn get_igv(&self) -> &Option<Decimal> {
         &self.igv
     }
 }
 
 impl DetalleIGVSetter for Detalle {
-    fn set_igv(&mut self, val: f64) {
+    fn set_igv(&mut self, val: Decimal) {
         self.igv = Some(val);
     }
 }
