+++ src/models/traits/detalle/precioreferencia.rs
@@ -1,21 +1,22 @@
 use crate::models::general::Detalle;
+use rust_decimal::Decimal;
 
 pub trait DetallePrecioReferenciaGetter {
-    fn get_precioreferencia(&self) -> &Option<f64>;
+    fn get_precioreferencia(&self) -> &Option<Decimal>;
 }
 
 pub trait DetallePrecioReferenciaSetter {
-    fn set_precioreferencia(&mut self, val: f64);
+    fn set_precioreferencia(&mut self, val: Decimal);
 }
 
 impl DetallePrecioReferenciaGetter for Detalle {
-    fn get_precioreferencia(&self) -> &Option<f64> {
+    fn get_precioreferencia(&self) -> &Option<Decimal> {
         &self.precio_referencia
     }
 }
 
 impl DetallePrecioReferenciaSetter for Detalle {
-    fn set_precioreferencia(&mut self, val: f64) {
+    fn set_precioreferencia(&mut self, val: Decimal) {
         self.precio_referencia = Some(val);
     }
 }
