+++ src/models/traits/detalle/totalimpuestos.rs
@@ -1,21 +1,22 @@
 use crate::models::general::Detalle;
+use rust_decimal::Decimal;
 
 pub trait DetalleTotalImpuestosGetter {
-    fn get_totalimpuestos(&self) -> &Option<f64>;
+    fn get_totalimpuestos(&self) -> &Option<Decimal>;
 }
 
 pub trait DetalleTotalImpuestosSetter {
-    fn set_totalimpuestos(&mut self, val: f64);
+    fn set_totalimpuestos(&mut self, val: Decimal);
 }
 
 impl DetalleTotalImpuestosGetter for Detalle {
-    fn get_totalimpuestos(&self) -> &Option<f64> {
+    fn get_totalimpuestos(&self) -> &Option<Decimal> {
         &self.total_impuestos
     }
 }
 
 impl DetalleTotalImpuestosSetter for Detalle {
-    fn set_totalimpuestos(&mut self, val: f64) {
+    fn set_totalimpuestos(&mut self, val: Decimal) {
         self.total_impuestos = Some(val);
     }
 }
